chore: refactor config package into separate files
To make the single responsibility principle clearer, the package is separated into different files, each with one responsibility

[#186464017](https://www.pivotaltracker.com/story/show/186464017)
