+++ internal/config/config.go
@@ -3,36 +3,22 @@ package config
 import (
 	"flag"
 	"fmt"
-	"os"
-	"regexp"
 	"strings"
-
-	"github.com/hashicorp/go-version"
 )
 
-const (
-	Usage = "cf upgrade-all-services <broker-name> [options]"
+// Config is the type that contains all the configuration data required to run the plugin
+type Config struct {
+	BrokerName        string
+	APIToken          string
+	APIEndpoint       string
+	SkipSSLValidation bool
+	HTTPLogging       bool
+	DryRun            bool
+	CheckUpToDate     bool
+	ParallelUpgrades  int
+}
 
-	parallelDefault     = 10
-	parallelFlag        = "parallel"
-	parallelDescription = "number of upgrades to run in parallel"
-
-	// Ideally we would have used "-v" as the flag as the CF CLI does,
-	// but unfortunately the CF CLI swallows this flag, and the value
-	// is not available to plugins
-	httpLoggingDefault     = false
-	httpLoggingFlag        = "loghttp"
-	httpLoggingDescription = "enable HTTP request logging"
-
-	dryRunDefault     = false
-	dryRunFlag        = "dry-run"
-	dryRunDescription = "print the service instances that would be upgraded"
-
-	checkUpToDateDefault     = false
-	checkUpToDateFlag        = "check-up-to-date"
-	checkUpToDateDescription = "checks and fails if any service instance is not up-to-date - implies a dry-run"
-)
-
+// ParseConfig combines and validates data from the command line and CLIConnection object
 func ParseConfig(conn CLIConnection, args []string) (Config, error) {
 	var cfg Config
 
@@ -42,6 +28,8 @@ func ParseConfig(conn CLIConnection, args []string) (Config, error) {
 	flagSet.BoolVar(&cfg.DryRun, dryRunFlag, dryRunDefault, dryRunDescription)
 	flagSet.BoolVar(&cfg.CheckUpToDate, checkUpToDateFlag, checkUpToDateDefault, checkUpToDateDescription)
 
+	// This ranges over a chain of functions, each of which performs a single action and may return an error.
+	// The chain breaks at the first error received. It arguably reads better than repetitive error handling logic.
 	for _, s := range []func() error{
 		func() error {
 			return validateLoginStatus(conn)
@@ -76,36 +64,7 @@ func ParseConfig(conn CLIConnection, args []string) (Config, error) {
 	return cfg, nil
 }
 
-type Config struct {
-	BrokerName        string
-	APIToken          string
-	APIEndpoint       string
-	SkipSSLValidation bool
-	HTTPLogging       bool
-	DryRun            bool
-	CheckUpToDate     bool
-	ParallelUpgrades  int
-}
-
-//go:generate go run github.com/maxbrunsfeld/counterfeiter/v6 -generate
-//counterfeiter:generate . CLIConnection
-type CLIConnection interface {
-	IsLoggedIn() (bool, error)
-	AccessToken() (string, error)
-	ApiVersion() (string, error)
-	ApiEndpoint() (string, error)
-	IsSSLDisabled() (bool, error)
-}
-
-func Options() map[string]string {
-	return map[string]string{
-		parallelFlag:      parallelDescription,
-		httpLoggingFlag:   httpLoggingDescription,
-		dryRunFlag:        dryRunDescription,
-		checkUpToDateFlag: checkUpToDateDescription,
-	}
-}
-
+// parseCommandLine reads the command line argument, with validation handled later by validation functions
 func parseCommandLine(flagSet *flag.FlagSet, args []string) (string, error) {
 	if len(args) == 0 {
 		printUsage()
@@ -124,6 +83,8 @@ func parseCommandLine(flagSet *flag.FlagSet, args []string) (string, error) {
 	return args[0], nil
 }
 
+// read calls a function (typically on the object that implements CLIConnection) and assuming
+// no error it stores it in the specified location. This arguably reads better than repetitive logic.
 func read[T any](desc string, get func() (T, error), set *T) error {
 	data, err := get()
 	if err != nil {
@@ -132,66 +93,4 @@ func read[T any](desc string, get func() (T, error), set *T) error {
 
 	*set = data
 	return nil
-}
-
-func validateLoginStatus(conn CLIConnection) error {
-	loggedIn, err := conn.IsLoggedIn()
-	switch {
-	case err != nil:
-		return fmt.Errorf("error getting login status: %w", err)
-	case !loggedIn:
-		return fmt.Errorf("you must authenticate with the cf cli before running this command")
-	default:
-		return nil
-	}
-}
-
-func validateAPIVersion(conn CLIConnection) error {
-	ver, err := conn.ApiVersion()
-	if err != nil {
-		return fmt.Errorf("error retrieving API version: %w", err)
-	}
-
-	var (
-		v3    = version.Must(version.NewVersion("3"))
-		v4    = version.Must(version.NewVersion("4"))
-		v2min = version.Must(version.NewVersion("2.164"))
-		v3min = version.Must(version.NewVersion("3.99"))
-	)
-
-	v, err := version.NewVersion(ver)
-	switch {
-	case err != nil:
-		return fmt.Errorf("error parsing API version: %w", err)
-	case v.GreaterThanOrEqual(v3min) && v.LessThan(v4):
-		return nil
-	case v.GreaterThanOrEqual(v2min) && v.LessThan(v3):
-		// There's a bug in CF CLI v6 where the API version is sometimes reported as v3 and sometimes as v2,
-		// depending on whether "cf login" of "cf api" was used. CAPI release 1.109.0 shipped with both
-		// API v3.99 and CF API v2.164, so if we have at least v2.164 then we know that v3.99 is also available
-		return nil
-	default:
-		return fmt.Errorf("plugin requires minimum API version %s or %s, got %q", v3min.String(), v2min.String(), v.String())
-	}
-}
-
-func validateParallelUpgrades(p int) error {
-	if p <= 0 || p > 100 {
-		printUsage()
-		return fmt.Errorf("number of parallel upgrades must be in the range of 1 to 100")
-	}
-	return nil
-}
-
-func validateBrokerName(name string) error {
-	if valid := regexp.MustCompile(`^[\w_.-]+$`).MatchString(name); !valid {
-		printUsage()
-		return fmt.Errorf("broker name contains invalid characters")
-	}
-
-	return nil
-}
-
-func printUsage() {
-	fmt.Fprintf(os.Stderr, "Usage: %s\n", Usage)
 }
